Refactor widgets into own package
