+++ bomi/device_managers/__init__.py
@@ -3,5 +3,3 @@ Device managers classes and their QWidget wrappers
 """
 from .protocols import SupportsStreaming, SupportsGetSensorMetadata, SupportsHasSensors
 from .yost_manager import YostDeviceManager
-from .yost_widget import YostWidget
-from .trigno_widget import TrignoClient, TrignoWidget
