+++ bomi/main.py
@@ -2,16 +2,16 @@ import PySide6.QtGui as qg
 import PySide6.QtWidgets as qw
 
 from bomi.device_managers.yost_manager import YostDeviceManager
-from bomi.device_managers.yost_widget import YostWidget
+from bomi.widgets.yost_widget import YostWidget
 
 from bomi.device_managers.qtm_manager import QtmDeviceManager
-from bomi.device_managers.qtm_widget import QtmWidget
+from bomi.widgets.qtm_widget import QtmWidget
 
-from bomi.start_react_widget import StartReactWidget
-from bomi.window_mixin import WindowMixin
-from bomi.base_widgets import wrap_gb
+from bomi.widgets.start_react_widget import StartReactWidget
+from bomi.widgets.window_mixin import WindowMixin
+from bomi.widgets.base_widgets import wrap_gb
 
-from bomi.device_managers.trigno_widget import TrignoWidget, TrignoClient
+from bomi.widgets.trigno_widget import TrignoWidget, TrignoClient
