+++ bomi/widgets/scope_widget.py
@@ -15,7 +15,7 @@ import PySide6.QtWidgets as qw
 from pyqtgraph.parametertree.parameterTypes import ActionParameter
 from pyqtgraph.parametertree.parameterTypes.basetypes import Parameter
 
-from bomi.base_widgets import TaskEvent, TaskDisplay, generate_edit_form
+from bomi.widgets.base_widgets import TaskEvent, TaskDisplay, generate_edit_form
 from bomi.datastructure import MultichannelBuffer, SubjectMetadata, Packet
 from bomi.device_managers.protocols import (
     SupportsStreaming,
