+++ bomi/widgets/yost_widget.py
@@ -9,7 +9,7 @@ import PySide6.QtWidgets as qw
 from PySide6.QtCore import Qt
 
 from bomi.datastructure import get_savedir
-from bomi.device_managers.table_model import (
+from bomi.widgets.table_model import (
     TableModel,
     ColumnProps,
     prop_getter,
@@ -17,8 +17,8 @@ from bomi.device_managers.table_model import (
     make_setter,
 )
 from bomi.device_managers.yost_manager import DeviceT, YostDeviceManager
-from bomi.scope_widget import ScopeWidget, ScopeConfig
-from bomi.window_mixin import WindowMixin
+from bomi.widgets.scope_widget import ScopeWidget, ScopeConfig
+from bomi.widgets.window_mixin import WindowMixin
 
 
 __all__ = ("YostWidget",)
